Detect raspberry pi
utils method for detecting raspi
split utils module into utils and math
Disable shadow window on raspi
