+++ arcade/emitter.py
@@ -6,7 +6,7 @@ over their lifetime
 import arcade
 from arcade.particle import Particle
 from typing import Optional, Callable, cast
-from arcade.utils import _Vec2
+from arcade.math import _Vec2
 from arcade.arcade_types import Point, Vector
